# openapi.yaml
openapi: 3.0.3
info:
  title: Car Rental Service 3.0
  description: |-
    Cервіс для прокату авто. Користувачі сервісу можуть бути двох рівнів – адміністратори та пасажири. Адміністратори можуть додавати та видаляти авто із системи, редагувати інформацію про авто. Пасажири можуть переглядати каталог та бронювати авто на певний час.
  version: 1.0.11
servers:
  - url: https://localhost:8080/api/v3
paths:
...
  /cars/{carId}:
    get:
      summary: Method for obtaining a car by Id
      tags:
        - cars
        - car
      operationId: getCarById
      parameters: 
        - name: carId
          in: path
          required: true
          description: Car Id
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: The request was completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        'default':
          description: Request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: The method of removing a car by Id
      tags: 
        - cars
        - car
      operationId: deleteCarById
      parameters: 
        - name: carId
          in: path
          required: true
          description: Car Id
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: The request was completed successfully
          content:
            application/json: {}
        'default':
          description: Request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Method of editing car data by Id 
      tags: 
        - cars
        - car
      operationId: updateCarById
      parameters: 
        - name: carId
          in: path
          required: true
          description: Car Id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Car'
      responses:
        '200':
          description: Update Succesfully
          content:
            application/json:
              schema:
                  items:
                    $ref: '#/components/schemas/Car'
        'default':
          description: Request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
#######################################################################################
...
components:
  schemas:
...
    Car:
      required:
        - status
        - brand
        - fuelСonsumption
      type: object
      properties:
        carId:
          type: integer
          format: int64
          example: 198772
        brand:
          type: string
          example: BMW
        model:
          type: string
          example: X5
        maxSpeed:
          type: integer
          format: int64
          example: 250
        yearProduction:
          type: integer
          format: int64
          example: 2019
        fuelСonsumption:
          type: number
          format: double
          example: 10.7
        seatsNumber:
          type: integer
          format: int64
          example: 5
        status:
          type: string
          description: Car Status 
          example: free
          enum:
            - free
            - reserve
        reservation:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
    Reservation:
      required:
        - reservId
        - startTime
        - endTime
      type: object
      properties:
        reservId:
          type: integer
          format: int64
          example: 54
        startTime:
          type: string
          format: date-time
          example: 2017-07-21T17:32:28Z
        endTime:
          type: string
          format: date-time
          example: 2017-07-21T17:32:28Z
...
    Error:
      required: 
        - code
        - message
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...